multi tiered test

+++ fireball_detection/dataset/scripts/dodgy.py
@@ -7,11 +7,15 @@ def main():
     if not GFO_FIXES_FOLDER.exists():
         os.mkdir(GFO_FIXES_FOLDER)
 
-    with open(Path(Path(__file__).parents[2], "data", "dodgy_fireballs.txt"), "r") as file:
+    with open(Path(Path(__file__).parents[3], "data", "dodgy_fireballs.txt"), "r") as file:
         dodgy_fireballs = [line.strip() for line in file.readlines()]
 
     for i in dodgy_fireballs:
         print(i)
+        csv_file = Path(GFO_PICKINGS, i + ".csv")
+        if not csv_file.exists():
+            print("does not exist")
+            continue
         shutil.copy(
             Path(GFO_PICKINGS, i + ".csv"),
             GFO_FIXES_FOLDER
